+++ SpeedyWeather/src/parameterizations/surface_fluxes/momentum.jl
@@ -27,7 +27,7 @@ initialize!(::SurfaceMomentumFlux, ::PrimitiveEquation) = nothing
 
 function variables(::SurfaceMomentumFlux)
     return (
-        ParameterizationVariable(:boundary_layer_drag, Grid2D(), desc = "Boundary layer drag coefficient", units = "1"),
+        ParameterizationVariable(:boundary_layer_drag_momentum, Grid2D(), desc = "Boundary layer drag coefficient for momentum", units = "1"),
         ParameterizationVariable(:surface_wind_speed, Grid2D(), desc = "Surface wind speed", units = "m/s"),
         ParameterizationVariable(:surface_air_density, Grid2D(), desc = "Surface air density", units = "kg/m^2"),
         ParameterizationVariable(:surface_air_temperature, Grid2D(), desc = "Surface air temperature", units = "K"),
@@ -47,7 +47,7 @@ end
     surface = model.geometry.nlayers
 
     # drag coefficient either from SurfaceMomentumFlux or from a central drag coefficient
-    d = vars.parameterizations.boundary_layer_drag[ij]
+    d = vars.parameterizations.boundary_layer_drag_momentum[ij]
     drag = ifelse(momentum_flux.use_boundary_layer_drag, d, land_fraction * drag_land + (1 - land_fraction) * drag_ocean)
 
     # Fortran SPEEDY documentation eq. 52, 53, accumulate fluxes with +=